requirements.txt
# OntologyOps Core Dependencies
rdflib>=7.0.0
networkx>=3.0
requests>=2.28.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Optional monitoring
prometheus-client>=0.16.0
